+++ qtrmrs/apps/users/models.py
@@ -7,6 +7,7 @@ from django.dispatch import receiver
 class User(AbstractUser):
     email = models.EmailField(unique=True)
     is_student = models.BooleanField(default=True)
+    is_email_verified = models.BooleanField(default=False)
     avatar = models.ImageField(upload_to='avatars/', null=True, blank=True)
 
     USERNAME_FIELD = 'email'